Fix image buffer index overflow and enforce ImageData validation
Compute draw_image src_idx in usize to prevent u32 wrapping on large
images. Promote ImageData::from_raw debug_assert to assert so buffer
size is validated in release builds.

Fixes review issues C4, H4.

+++ crates/aurora_render/src/canvas.rs
@@ -694,7 +694,7 @@ impl<'a> Canvas<'a> {
                     continue;
                 }
 
-                let src_idx = ((sy * img_width + sx) * 4) as usize;
+                let src_idx = ((sy as usize) * (img_width as usize) + (sx as usize)) * 4;
                 if src_idx + 3 >= pixels.len() {
                     continue;
                 }
